Worked on edit for sale and sale detail Worked on making sale delete cascade Worked on adding environment values for munshiana.

+++ src/app/components/sales-list/sales-list.component.ts
@@ -1,5 +1,7 @@
 import { Sale } from '../../shared/models/sale';
 import { SaleApiService } from '../../shared/services/sales.service';
+import { SaleDetailApiService } from '../../shared/services/saleDetail.service';
+import { ApiService as SupplierService } from '../../shared/services/suppliers.service';
 import { Component, ViewChild, OnInit } from '@angular/core';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
@@ -14,12 +16,17 @@ export class SalesListComponent implements OnInit {
   SaleData: any = [];
   dataSource: MatTableDataSource<Sale>;
   @ViewChild(MatPaginator, { static: false }) paginator: MatPaginator;
-  displayedColumns: string[] = ['_id', 'builtyNumber', 'quantity', 'amount', 'action'];
+  displayedColumns: string[] = ['_id', 'builtyNumber', 'truckNumber', 'truckRent', 'supplier', 'action'];
 
-  constructor(private saleApi: SaleApiService) {
+  constructor(private saleApi: SaleApiService, private supplierApi: SupplierService, private saleDetailApi: SaleDetailApiService) {
     this.saleApi.GetSales().subscribe(data => {
       this.SaleData = data;
-      console.log(this.SaleData);
+      this.SaleData.forEach(element => {
+        this.supplierApi.GetSupplier(element.supplier).subscribe((supplier) => {
+          console.log(supplier);
+          element.supplier = supplier.name;
+        });
+      });
       this.dataSource = new MatTableDataSource<Sale>(this.SaleData);
       setTimeout(() => {
         this.dataSource.paginator = this.paginator;
@@ -27,13 +34,30 @@ export class SalesListComponent implements OnInit {
     });
   }
 
-  ngOnInit() { }
+  ngOnInit() {
+
+  }
+
+  public doFilter = (value: string) => {
+    this.dataSource.filter = value.trim().toLocaleLowerCase();
+  }
+  getSupplierName(id) {
+    this.supplierApi.GetSupplier(id).subscribe((data) => {
+      console.log(data.name);
+      return data.name;
+    });
+  }
 
   deleteSale(index: number, e) {
-    if (window.confirm('Are you sure')) {
+    if (window.confirm('Are you sure, as this will delete all sale and detail information as well.')) {
       const data = this.dataSource.data;
       data.splice((this.paginator.pageIndex * this.paginator.pageSize) + index, 1);
       this.dataSource.data = data;
+      this.saleApi.GetSale(e._id).subscribe((sale) => {
+        sale.saleDetails.forEach(element => {
+          this.saleDetailApi.DeleteSaleDetail(element);
+        });
+      });
       this.saleApi.DeleteSale(e._id).subscribe();
     }
   }
